+++ client/src/app/app-routing.module.ts
@@ -2,8 +2,8 @@ import { APP_BASE_HREF } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
-import { QuizComponent } from './components/quiz/quiz.component';
-import { ScoreboardComponent } from './components/scoreboard/scoreboard.component';
+import { QuizComponent } from './components/quiz/quiz/quiz.component';
+import { ScoreboardComponent } from './components/scoreboard/scoreboard/scoreboard.component';
 
 const routes: Routes = [
   { path: 'quiz/:game', component: QuizComponent },
