+++ client/src/app/components/nav-bar/nav-bar.component.html
@@ -1,4 +1,4 @@
 <div class="navigation">
-    <a [routerLink]="['/quiz']" class="active">Take Quiz</a>
-    <a [routerLink]="['/scoreboard']">View Scoreboard</a>
+    <a [routerLink]="['/quiz']" [class]="(currentPage == 'quiz') ? 'active' : ''">Take Quiz</a>
+    <a [routerLink]="['/scoreboard']" [class]="(currentPage == 'scoreboard') ? 'active' : ''">View Scoreboard</a>
 </div>
