+++ client/src/app/components/quiz/quiz/quiz.component.html
@@ -3,7 +3,7 @@
         <thead>
             <tr>
                 <td id="navigation" colspan="3">
-                    <app-nav-bar></app-nav-bar>
+                    <app-nav-bar [currentPage]="'quiz'"></app-nav-bar>
                 </td>
             </tr>
             <tr id="nameAndScore">
