+++ client/src/app/components/scoreboard/scoreboard-row.css
