+++ client/src/app/components/scoreboard/scoreboard-row/scoreboard-row.component.spec.ts
