+++ client/src/app/components/scoreboard/scoreboard-row/scoreboard-row.component.ts
@@ -4,7 +4,7 @@ import { PlayerData } from 'src/app/model/player-data';
 @Component({
   selector: '[app-scoreboard-row]',
   templateUrl: './scoreboard-row.component.html',
-  styleUrls: ['../shared/scoreboard-row.css', './scoreboard-row.component.css']
+  styleUrls: ['../scoreboard-row.css', './scoreboard-row.component.css']
 })
 export class ScoreboardRowComponent implements OnInit {
   @Input() playerData! : PlayerData;
