+++ client/src/app/components/scoreboard/scoreboard/scoreboard.component.html
@@ -2,7 +2,7 @@
     <thead>
         <tr>
             <td id="navigation" colspan="3">
-                <app-nav-bar></app-nav-bar>
+                <app-nav-bar [currentPage]="'scoreboard'"></app-nav-bar>
             </td>
         </tr>
         <tr>
