--- client/src/app/components/shared/quiz.css
@@ -1,51 +1,0 @@
-* {
-  padding: 0px;
-  margin: 0px;
-}
-
-#big {
-  margin: 0px auto;
-}
-
-#head, #done {
-  margin: 30px 20px;
-  font-size: 20px;
-}
-
-#name {
-  font-size: 16px;
-}
-
-table {
-  border-collapse: collapse;
-  margin: 0px auto;
-}
-
-#nameAndScore td {
-  padding: 20px 5px;
-  margin: 0px;
-}
-
-td:first-child {
-  width: 300px;
-}
-
-#done {
-  height: 50px;
-  width: 100px;
-  border-radius: 5px;
-}
-
-#scoreLabel, #fractionLabel {
-  font-size: 20px;
-  font-family: sans-serif;
-  text-align: center;
-}
-
-.correct, .complete {
-  background-color: lightgreen;
-}
-
-.incorrect, .incomplete {
-  background-color: #faa;
-}
