+++ client/src/styles.css
@@ -1,1 +1,52 @@
 /* You can add global styles to this file, and also import other style files */
+* {
+  padding: 0px;
+  margin: 0px;
+}
+
+#big {
+  margin: 0px auto;
+}
+
+#head, #done {
+  margin: 30px 20px;
+  font-size: 20px;
+}
+
+#name {
+  font-size: 16px;
+}
+
+table {
+  border-collapse: collapse;
+  margin: 0px auto;
+}
+
+#nameAndScore td {
+  padding: 20px 5px;
+  margin: 0px;
+}
+
+td:first-child {
+  width: 300px;
+}
+
+#done {
+  height: 50px;
+  width: 100px;
+  border-radius: 5px;
+}
+
+#scoreLabel, #fractionLabel {
+  font-size: 20px;
+  font-family: sans-serif;
+  text-align: center;
+}
+
+.correct, .complete {
+  background-color: lightgreen;
+}
+
+.incorrect, .incomplete {
+  background-color: #faa;
+}
